MockController now works with any Lock.
I could not figure out the type magic to get IO-based MockController to
work with a Monad-based Lock interface, so in order to make
MockController work with abstract Locks, I had to parameterize Lock on
MonadIO, rather than Monad. That's unfortunate in that it's less pure,
though in theory all useful locks will be IO-based, of course.