Add "lock at <time>" command to the state machine.
When an unlock command is executed by the Mellon state machine, it
unlocks the lock device and creates a lock command that will run in the
future, at the date specified by the unlock command. Obviously, in order
to keep the state machine consistent, this "future lock" must itself be
executed by the state machine, just as a user-initiated command would
be. (Note that the Mellon state machine does not specify how these
"future locks" are scheduled or sent to the state machine when
triggered; that is the responsibility of a controller implementation.)

Prior to this change, there was only one kind of "lock" command:
effectively, a "lock now" command, which was used both by the user in
order to override any current "unlock until <time>" commands in effect;
and by a Mellon controller, upon the expiration of a previously-executed
"unlock until <time>" command.

In a real-world controller implementation (e.g., ThreadedController),
most likely the future lock will be scheduled on and run from a
different thread of control than the state machine thread of control.
Given this concurrency model and just the one flavor of lock
command ("lock now"), there exists the possibility of a race. For
example, imagine the following scenario:

1. An "unlock until midnight" command is executed by the controller. The
lock device is unlocked, and a lock command is scheduled to run at
midnight on a new background thread.

2. At midnight, the background thread that was created by the earlier
"unlock until midnight" command sends its "lock now" command to the
controller thread.

3. Also at midnight, a user process sends an "unlock until midnight+1"
command to the controller thread.

Ideally, what should happen in this scenario? The lock device should
remain unlocked until midnight+1, of course, even if it briefly locks in
between the sequence of unlock commands.

However, if the user's "unlock until midnight+1" command wins the race
to the controller, followed immediately thereafter by the "lock now"
command from the previous (and now overridden) "unlock until midnight"
command, then effectively the "unlock until midnight+1" command will be
undone by the earlier, expired unlock command, and the device will lock
approximately at midnight, rather than remaining unlocked until
midnight+1. That's not what we want.

The most elegant way to eliminate this race condition is to tag a
scheduled lock command with the expiration date of the unlock command
from which it was generated. Then, by the time the controller processes
the scheduled lock command, if there is a new unlock command already in
progress with a later expiration than the scheduled lock command, the
controller can simply ignore the lock command. (When the new "unlock at
<time>" command expires, its scheduled "lock at <time>" command will
fire, and the device will still lock as expected.)

Accordingly, this change adds a new "lock at <time>" command to the
state machine. Note that this new "lock at <time>" command does not
replace the existing "lock now" command; the latter command still has a
purpose as a way for the user to request that the device is locked
immediately, irrespective of any in-progress unlock commands. However,
in the interest of keeping the user's view of the Mellon state machine
as simple as possible, controller implementations should take care not
to expose the "lock at <time>" state machine command to the user.

Therefore, from the user's perspective, this commit introduces no
changes to the Mellon state machine model: it is still only the "lock
now" and "unlock until <time>" commands that are available to the user.