runStateMachine: ignore locks when state is Locked
Previously, runStateMachine would issue a lock command even when its
current state was already Locked. Originally I figured this would be a
good way to sync up the controller implementation and the abstract state
machine, but since the controller implementation a) most likely has
direct access to the lock device itself and b) can initialize the state
machine in any state it likes, these redundant lock commands in the
state machine are unnecessary, and may even be the cause of confusion if
they show up in log files, etc.