Remove MonadLock IO instance.
Note that I've replaced one hack with another. In
runConcurrentControllerT, in order to wedge MonadLock into the state
machine thread, you have to "run" a MonadLock computation. Right now
there's only MockLockT, so that's what I've used as a first step.
Eventually, of course, the caller of runConcurrentControllerT will want
to choose the MonadLock implementation used by the embedded
StateMachineT, and may want to see the unwrapped MonadLock's value, as
well (e.g., for testing).

This will require a rethink of the ConcurrentController interface as
right now the caller never sees the StateMachineT at all.