Break ConcurrentController into constituent parts.
Caller now creates a CC (i.e., an MVar Cmd), then runs both the
concurrent state machine and the concurrent controller. This means the
caller can decide a) which MonadLock implementation to use with the
state machine and b) which one to fork and which one to run in the
current thread (or fork both).