Add a Halt state to StateMachineF.
And a corresponding QuitCmd for controllers to implement (via 'halt').

This new state is needed so that, if the caller forks
runConcurrentController and does runConcurrentStateMachine in the
foreground thread, the latter doesn't hang in takeMVar on the shared
MVar when the former returns.

Note that I had to hack up the function signatures of
runConcurrentControllerT and runConcurrentStateMachine to make this
work. In runConcurrentStateMachine the implementation of Halt needs to
'return' a value, and runConcurrentControllerT must put QuitCmd in the
shared MVar in order to tell the state machine to halt. That means we
need a default value to return in both functions, so for now it's always
'()'. Eventually the signatures should be changed so that the caller
must pass a "finish" function that provides the return value, or perhaps
runConcurrentStateMachine could pass a computed value back to
runConcurrentControllerT via another MVar. In that way the threads could
coordinate upon quit.