Sync concurrent controller and state machine.
Now, when runConcurrentControllerT is finished running the control
program, runConcurrentControllerT will synchronize with
runConcurrentStateMachine to ensure that both have quit. Additionally,
runConcurrentStateMachine will send a user-supplied return value to
runConcurrentControllerT, and both will return this value as the result
of the computed Monad.

Currently, the user-supplied return value is simply of parameterized
type 'a'. It should probably be a function of some kind; however, it's
not clear what that function should take as an argument. Need to think
about this some more.