Move state machine event loop out of StateMachineT
The *mechanism of how the state machine waits for events* is independent
of the *mechanism of transitioning from state to state*, which is what
StateMachineT is really about. Therefore I've removed the waitForCmd and
halt verbs from the StateMachineT eDSL, and I've also removed the hack-y
QuitCmd state machine command.

All of this stuff is important, of course, and has simply moved
elsewhere (currently into the guts of each Controller implementation),
and it could even be an eDSL/free monad of its own right, but the point
is that the StateMachineT free monad is much cleaner if it doesn't
attempt to deal with these things that are really a separate concern.

This enables us to tailor the event loop mechanism and the business of
halting the state machine into the implementation of the controller,
which is where it belongs.