runConcurrent* now returns a computed value.
Specifically, the computation provided to runConcurrentControllerT is
evaluated and its value is returned both by runConcurrentControllerT,
and by runConcurrentStateMachine. In the latter case, the value is
communicated to runConcurrentStateMachine by runConcurrentControllerT
via the "quit" MVar they share. This allows the caller of
runConcurrentStateMachine, which is on a different thread than
runConcurrentControllerT, to see the value returned by the concurrent
controller computation.