Step one of MonadController refactoring.
ConcurrentControllerT doesn't do everything it's supposed to do yet, but
it does work with the new MonadController interface, which replaces the
FreeT implementation of controllers. Because the controller DSL is so
simple (just lockNow and unlockUntil), it's not clear that the overhead
and conceptual complexity of using FreeT is worth it, versus just making
MonadController a Monad instance and then defining all the necessary
Monad stuff to create each MonadController instance.

Making ConcurrentControllerT a Monad is easy because it's just a wrapper
around ReaderT.

What's left to do in order to get back to full functionality is to run
the state machine from ConcurrentControllerT's MonadController instance.
The current plan is to change the previous implementation, in which
lockNow and unlockUntil are just relatively straightforward functions
using MVars to communicate with a concurrently-running
runConcurrentStateMachine thread, where all the magic happens; to
a (hopefully) simpler implementation where lockNow and unlockUntil run
the state machine in the current thread, and the MVar is used as a mutex
around the state machine (which really needs to run in mutual exclusion
due to the nature of the state machine transitions). This has some
advantages and disadvantages, but I'll save the discussion of those for
the future check-in which implements the new mechanism.