Simpler ConcurrentControllerT API (with caveats).
This version is much simpler, both from an implementation perspective,
and a use perspective. For one thing, each controller thread now runs
the state machine in its own thread, whereas in the previous
implementation, a "controller thread" just passed messages to a single
state machine thread. In the new implementation, the StateMachineT API
is protected by a mutex in an MVar shared by all controller threads.
When the user calls lockNow/unlockUntil on the controller, the calling
thread will block until it acquires the mutex and then iterate the state
machine for the managed locked device. This eliminates a whole class of
sticky problems inherent in the previous design, one of which is how to
manage "child" threads and perform proper cleanup when the state machine
runs in its own thread (a problem that was never properly solved in the
original implementation!).

The new code is much simpler and shorter, as all of the message-passing
and "quit" synchronization code has been eliminated. The new
StateMachineF EDSL implementation is absolutely trivial and maps
directly on to ConcurrentControllerT or MonadLock primitives.

There are two drawbacks to this approach:

1. The calling thread may now block longer than it used to in the
previous design, when a lock or unlock operation is performed.
Previously, the calling thread passed the command message and then
continued executing, without waiting for the device to lock or
unlock. (I assume here that passing a message would usually be faster
than synchronously waiting for the lock event, since they are both based
on a similar MVar mechanism.)

Basically, in the new design, locking and unlocking is synchronous. This
actually may not be a drawback, as such. The synchronous approach paves
the way for safer/more robust design changes in the future, as now we
could return a "success" value from the MonadLock lock/unlock and
MonadController lockNow/unlockUntil functions, for MonadLock
implementations that may fail for some reason. (The StateMachine
implementation would also need to be modified to accommodate the
possibility of failure, however.)

Also, if blocking a thread is a concern, this new design does not
preclude layering an asynchronous design on top -- you could, for
example, write an AsyncController monad that wraps ConcurrentControllerT
with the old message-passing behavior, so that the caller of
lockNow/unlockUntil in a AsyncController computation would not have to
wait for the device to complete the command.

2. ... It doesn't actually work yet. There remains the sticky problem of
how to "run" the MonadLock locking/unlocking from inside the
ConcurrentControllerT. Currently, the implementation does not handle
scheduled lock events correctly. (It does launch a thread which sleeps for
the appropriate amount of time until the unlock expires; however, upon
wake-up, no LockCmd is executed and the thread simply exits silently.)

The reason this doesn't work yet is fairly complicated and is a good
example of why composing monads is so tricky. The MonadLock design is
based on the assumptions of the previous ConcurrentControllerT design,
namely that there would be one "master" state machine thread responsible
for locking and unlocking. Therefore, in the previous design, when the
program forks the state machine/MonadLock thread, it wraps the state
machine computation with a "runLock" function that implements the
MonadLock functionality invoked inside the state machine computation.
The key is that there is a *single MonadLock computation*, and as long
as you can run in the IO monad (i.e., pass a message via a shared MVar),
you can tell the MonadLock computation what to do. Essentially, there is
nothing "controller-y" (monad-wise) about controller threads in the
previous design; they're just IO computations and therefore can be used
just about anywhere that's not pure.

In the new ConcurrentControllerT design, we have eliminated message
passing between user threads and a single MonadLock computation. The
only way in the new system to interact with the MonadLock is by running
runConcurrentControllerT on a ConcurrentControllerT computation. At the
moment, there is no way to run a "runLock" function from inside the
ConcurrentControllerT computation, and that is the problem. The monad
wrapped by ConcurrentControllerT is required to be an instance of
MonadLock, of course, so that we can call lock and unlock from inside
the ConcurrentControllerT computation; but that computation only knows
about the MonadLock *interface*, not the particular
MonadLock *implementation*.

In the current implementation, the *caller* of runConcurrentControllerT
is responsible for running the MonadLock by wrapping the
ConcurrentControllerT around it and then calling a "runLock" function
when running the controller, as well. So what we need is some way to
move the "runLock" function *inside* the ConcurrentControllerT
computation, so that, when the time comes to run a scheduled lock in
response to an expiring unlock command, that thread can run its
computation inside the MonadLock.

So, in summary, what this new design has done is to *distribute the
MonadLock computation across multiple threads*, rather
than *centralizing it in a single thread* as the previous design did.
This makes managing the controller threads trivial and eliminates the
need for message passing in the controller, but it does mean that we can
no longer communicate with the MonadLock computation with just plain IO.
Every controller thread now needs to know how to run a MonadLock
computation.

This also means that, if we need to manage some kind of shared lock
state in the MonadLock implementation (say, the list of MockLockEvent
events in MockLockT), then we have effectively just moved the
synchronization complexity from ConcurrentControllerT into the MonadLock
implementation, instead. MonadLock is conceptually a much simpler
interface than a controller, because MonadLock doesn't need to interact
with the state machine, so I think it's probably much easier to add a
bit of complexity to the MonadLock implementation than to try to
coordinate worker threads with a "master" thread, as the previous design
did.

(One way in which MonadLock synchronization is easier than controller
synchronization is that MonadLock implementations will not necessarily
have to deal with re-entrancy -- the design of ConcurrentControllerT
ensures that MonadLock's lock and unlock functions are only ever called
when inside the ConcurrentControllerT mutex. As long as there is a
one-to-one relationship between a MonadLock instance and a shared mutex,
there is no possibility that the MonadLock functions will be invoked
concurrently.)