Proof-of-concept with MockLock.
Stuff a MockLock into ConcurrentController, and then change
ConcurrentControllerT so that it does a "runMockLockT" *inside* the
controller implementation whenever it needs to lock or unlock the
device. This allows us to fork a thread that can
runConcurrentControllerT from IO () and still know how to lock and
unlock the MockLock.

This works great! Now just to make it work with a parameterized MonadLock
constraint so that it works with all MonadLock instances....