Working, if rather clumsy, runLockT implementation
Slight twist on my earlier plan: rather than passing a "runLock"
function into the ConcurrentControllerT, use the
MonadRandom/RandomGen/runRandT approach and create a generic
LockT/runLockT monad that wraps a LockDevice class. Then, in the
ConcurrentControllerT implementation, we can just grab the LockDevice
from the controller state and runLockT it.

(Actually, we should probably runLockT on the runConcurrentControllerT
so that ConcurrentControllerT can just use the MonadLock 'lock' and
'unlock' functions directly, but that will require a bit more type
signature trickery; save that for the next check-in.)