Force evaluation of new lock state.
In ConcurrentControllerT, we want to be sure that the lock/unlock
command is executed stricly by the state machine when the program
requests it, rather than lazily the next time someone tries to read the
MVar where the new state is stored, which could be much, much later!

Because of the many layers (ReaderT, FreeT, MVar, IO, lazy evaluation),
I'm not certain this is strictly (pun intended) needed, but it won't
hurt, at least.