Mimic Controller.State in Server.
Prior to this change, ServerState was a newtype wrapper for
Controller.State, but its ToJSON instance had to be written by hand
because Controller does not depend on Aeson, so there cannot be
ToJSON/FromJSON instances for Controller.State without creating an
orphan instances in Server. Furthermore, there was no FromJSON instance
because writing a FromJSON instance by hand for a sum type is pretty
ugly.

Anyway, it would be convenient to have safe, automatically-derived
ToJSON/FromJSON instances for the Server state, so rather than do the
straightforward/lazy thing and make Aeson a dependency of Controller,
this commit introduces a Server.State type which simply mimics
one-to-one the values of Controller.State. Then we write the
ToJSON/FromJSON instances in Server, where we already depend on Aeson.