Simplify the mellon-pi user interface.
Modulo concurrency issues, PiLock is an extremely simple mechanism and
probably doesn't have a lot of use as a standalone interface, so let's
not expose it. That means we don't need to expose a "withPiLock"
function to the user and hope that she uses it correctly; each *Server
function will just set up GPIO and the specified pin properly before
creating the ConcurrentControllerT monad and the mellon server. The user
of the API never sees the lock that gets created.

(This also means that if we decide later to move all of the GPIO stuff
to a separate thread, for exception and thread safety, that the *Server
functions can take care of that, too, and not burden the user with
getting it right.)