Drop HPi, use gpio instead.
This change makes it possible to use any platform with GPIO support,
rather than just RPi/RPi 2. (Note that, currently, only the Linux sysfs
GPIO interface is supported, but it should be really easy to add other
targets when the gpio package supports them.)

The new package, mellon-gpio, is effectively the same as the old
one (mellon-pi). The only functional difference is that the executable
uses GPIO pin 65 rather than the old pin 3. This is because, on the
BeagleBone Black with CBB Relay cape that I'm using to test all of this
stuff, pin 65 drives the first relay on this board.