First steps towards a simpler mellon.
This change is the first in a series of planned changes to make
mellon (specifically, the core mellon package) simpler. These changes
will be rolled out side-by-side with the existing, functioning mellon
package in a new package, mellon-core. These two packages will continue
to exist until I'm ready to switch everything over to mellon-core and
remove the mellon package.

In this first change, I've rewritten the mellon state machine as just a
simple function; i.e., it's no longer a monad, nor a DSL. Originally, I
made it a monad because I wrote it as a DSL which needed to call back
into the controller to get the controller to do the right thing on each
state transition. For several of the transitions, there are multiple
actions required of the controller on behalf of the state machine (e.g.,
unlock the device and schedule a future lock event), but I realized
while re-writing this version that, because the actions performed by the
controller don't return anything to the state machine (i.e., the
transitional communication is one-way, from state machine to
controller), these multiple actions could combined into a compound
command and, therefore, need not be monadic.

With this change, the new controller design (not yet implemented) need
not be an instance of a state machine monad; the new controller will
simply iterate over the machine's states purely and perform the
action(s) specified by the machine's output on each transition.

Note that the new implementation of the state machine in mellon-core
behaves (in an abstract sense, anyway) exactly the same as the original
implementation in the mellon package. There are no behavioral changes.