Unlock times in the past do *not* indefinitely unlock the device.
This was my original intention, but the current implementation of the
controller does not, in fact, do this. Instead, it will very briefly
unlock the device and then lock it again; the amount of time this takes
depends on how quickly the lock thread can be scheduled.

I now consider both the original intended behavior (unlock indefinitely)
and the current behavior ("glitch" the lock) to be bugs. This is called
out explicitly now in the documentation for the async controller's
'unlockController' action.

If we want to support indefinite unlocks, we should provide a type
wrapper around the unlock time argument (a la Maybe) to indicate that
the lifetime of the request is either definite, or indefinite.

However, even if we do this, we still need a more considered approach to
dealing with unlock times in the past. A legitimate reason for a past
unlock date might be a command whose delivery was delayed by the
network. An illegitimate reason might be a "fuzzing" attempt. In both
cases, the appropriate response is probably to drop the request and to
indicate to the user that the request was ignored.

However, this will complicate both the controller's return type, and the
implementation of the controller (and potentially the state machine,
depending on where this situation is handled).