A proof-of-concept controller which ignores short/past unlock requests.
In this implementation, if the requested unlock date is within 1 second
of the time at which the request arrives at the controller, the
controller will simply advance the state machine to the next
state (locked) as if the unlock expired immediately. In addition to
ensuring that the device won't glitch, this change also takes care of
delayed/nonsensical unlock requests.

This works, but the delta should be configurable when the controller is
created.