Add a new output to the state machine.
The new output indicates to the controller that the currently-scheduled
lock request must be rescheduled. This makes apparent to the controller
the difference between the current unlock period being extended and a
locked -> unlock transition. In the former case, the controller doesn't
have to worry about a rapid locked->unlocked->locked glitch transition,
as it knows that the device is already unlocked.

Previously, both of these scenarios were aliased onto a single
'OutputUnlock' output value.