Make sure haskellPackages.mellon-* doesn't run unsafe tests.
When upstream packages use this overlay, we want to make sure that
tests that aren't CI-friendly don't cause build failures, but at the
same time, we want to run these tests on our own CI builds for this
repo. Therefore, we set haskellPackages.mellon-* to skip those tests,
and add a new package, haskellPackages.mellon-*-all-tests, which is
what the jobsets for this repo will build.